updating build to 2024.4
fixed NULL acceleration value - should be zero. removed unused variable for vision measurements

fixing loop comparisons-using size_t to cast loop comparison variable instead of int, vector size is unsigned

removed unused variables

fixed inializer list

fixed function to clear trajectory vector

initializer list error

elevarm warning fix - initialize order

fixing simple +/-

initalizer lists

signed-unsigned int comparison warning

signed-unsigned int comparison warning

moving robot pointer to protected to fix inializer order warning

+++ src/main/cpp/auto/ValorAuto.cpp
@@ -65,7 +65,7 @@ std::vector<std::string> listDirectory(std::string path_name){
 frc2::SequentialCommandGroup* ValorAuto::compileCommands(std::vector<ValorAutoAction> actions) {
     frc2::SequentialCommandGroup * commandGroup = new frc2::SequentialCommandGroup();
 
-    for (int i = 0; i < actions.size(); i ++) {
+    for (size_t i = 0; i < actions.size(); i ++) {
         ValorAutoAction & action = actions[i];
 
         // Error encountered: stop trying to compile and output error
